requirements.txt
torch
torchvision
numpy
scikit-learn
Pillow
PyYAML
timm
transformers
pandas
openpyxl
einops
opencv-python
